修复产品系统原生系统互切几率失败的问题

+++ drivers/misc_for_core/lidbg_misc.c
@@ -236,49 +236,40 @@ void cb_kv_lidbg_origin_system(char *key, char *value)
     LIDBG_WARN("\n\n<------------------system switch -%s----------->\n\n", value[0] == '1' ? "origin system" : "flyaudio system");
     if(value && *value == '1')//origin
     {
-        lidbg_toast_show("\"swtich to [origin system]\"", 0);
-        lidbg_shell_cmd("mkdir -p /flyapdata/.out/temp");
-        lidbg_shell_cmd("mkdir -p /flysystem/.out/temp");
+        lidbg_shell_cmd("mount -o remount /system 2> /dev/lidbg_msg");
+        lidbg_shell_cmd("mkdir -p /flyapdata/.out/temp 2> /dev/lidbg_msg");
+        lidbg_shell_cmd("mkdir -p /flysystem/.out/temp 2> /dev/lidbg_msg");
 
-        fs_remount_system();
+        lidbg_shell_cmd("cp  /flyapdata/app/ESFileExplorer.apk /system/app 2> /dev/lidbg_msg");
+        lidbg_shell_cmd("cp /flysystem/app/FlyBootService.apk /system/app 2> /dev/lidbg_msg");
 
-        lidbg_shell_cmd("cp -rf "ORIGIN_APP_PATH"* /system/priv-app/ &");
-        lidbg_shell_cmd("cp -rf "ORIGIN_TMP_PATH"NfcNci.apk /system/app/ &");
-        lidbg_shell_cmd("cp -r /flyapdata/app/ESFileExplorer.apk /system/app/ &");
-        lidbg_shell_cmd("cp -r /flysystem/app/FlyBootService.apk /system/app/ &");
-        lidbg_shell_cmd("cp -r /flysystem/app/FastBoot.apk /system/app/ &");
-        ssleep(15);//give sometime for cp
+        lidbg_shell_cmd("cp  "ORIGIN_APP_PATH"* /system/priv-app 2> /dev/lidbg_msg");
+        lidbg_shell_cmd("cp "ORIGIN_TMP_PATH"NfcNci.apk /system/app 2> /dev/lidbg_msg");
 
-        lidbg_shell_cmd("mv /flyapdata/* /flyapdata/.out/temp &");
-        lidbg_shell_cmd("mv /flysystem/* /flysystem/.out/temp &");
-
-        lidbg_shell_cmd("rm -r /system/priv-app/Launcher3.apk");
-        lidbg_shell_cmd("chmod 777 /system/app/F*");
-        lidbg_shell_cmd("chmod 777 /system/app/ESFileExplorer.apk");
-        lidbg_shell_cmd("chmod 777 /system/app/NfcNci.apk");
+        lidbg_shell_cmd("rm  /system/priv-app/Launcher3.apk 2> /dev/lidbg_msg");
+        lidbg_shell_cmd("chmod 777 /system/app/F* 2> /dev/lidbg_msg");
+        lidbg_shell_cmd("chmod 777 /system/app/ESFileExplorer.apk 2> /dev/lidbg_msg");
+        lidbg_shell_cmd("chmod 777 /system/app/NfcNci.apk 2> /dev/lidbg_msg");
+        lidbg_shell_cmd("cp /flysystem/app/FastBoot.apk /system/app 2> /dev/lidbg_msg");
+        while(!fs_is_file_exist("/system/app/FastBoot.apk"))
+            ssleep(2);
+        lidbg_shell_cmd("mv /flyapdata/* /flyapdata/.out/temp 2> /dev/lidbg_msg");
+        lidbg_shell_cmd("mv /flysystem/* /flysystem/.out/temp 2> /dev/lidbg_msg");
         goto suc;
     }
     else   if(value && *value == '2')//flyaudio
     {
-        lidbg_toast_show("\"swtich to [flyaudio system]\"", 0);
-
-        lidbg_shell_cmd("cp -rf  /flysystem/.out/temp/app/.sys-app1/* /system/priv-app &");
-
-        ssleep(10);//give sometime for cp
-        if(!fs_is_file_exist("/system/priv-app/Launcher3.apk"))
-        {
-            lidbg_shell_cmd("cp -r /flysystem/.out/temp/app/.sys-app1/Launcher3.apk /system/priv-app/ &");
-            ssleep(5);
-        }
-        lidbg_shell_cmd("mv /flysystem/.out/temp/* /flysystem &");
-        lidbg_shell_cmd("mv /flyapdata/.out/temp/* /flyapdata  &");
-        ssleep(5);//give sometime for cp
-
-        lidbg_shell_cmd("rm -r /system/app/ESFileExplorer.apk");
-        lidbg_shell_cmd("rm -r /system/app/FlyBootService.apk");
-        lidbg_shell_cmd("rm -r /system/app/FastBoot.apk");
-        lidbg_shell_cmd("rm -r /system/app/NfcNci.apk ");
-        lidbg_shell_cmd("rm -r /system/priv-app/Launcher2.apk");
+        lidbg_shell_cmd("mount -o remount /system 2> /dev/lidbg_msg");
+        lidbg_shell_cmd("mv /flysystem/.out/temp/* /flysystem");
+        lidbg_shell_cmd("mv /flyapdata/.out/temp/* /flyapdata");
+        lidbg_shell_cmd("cp  /flysystem/app/.sys-app1/* /system/priv-app 2> /dev/lidbg_msg");
+        lidbg_shell_cmd("rm /system/app/ESFileExplorer.apk 2> /dev/lidbg_msg");
+        lidbg_shell_cmd("rm /system/app/NfcNci.apk 2> /dev/lidbg_msg");
+        lidbg_shell_cmd("rm /system/priv-app/Launcher2.apk 2> /dev/lidbg_msg");
+        lidbg_shell_cmd("rm /system/app/FlyBootService.apk 2> /dev/lidbg_msg");
+        lidbg_shell_cmd("rm /system/app/FastBoot.apk 2> /dev/lidbg_msg");
+        while(!fs_is_file_exist("/system/priv-app/Launcher3.apk"))
+            ssleep(2);
         goto suc;
     }
 
@@ -308,7 +299,7 @@ int misc_init(void *data)
     lidbg_shell_cmd("mkdir  "ORIGIN_APP_PATH);
     lidbg_shell_cmd("mkdir  "ORIGIN_TMP_PATH);
 
-    if(!fs_is_file_exist(ORIGIN_APP_PATH"SystemUI.apk"))
+    if(!fs_is_file_exist(ORIGIN_APP_PATH"SystemUI.apk") && (fs_is_file_exist(FLY_MODE_FILE)))
     {
         lidbg_shell_cmd("mv /system/priv-app/SystemUI.apk "ORIGIN_APP_PATH"SystemUI.apk" );
         lidbg_shell_cmd("mv /system/priv-app/Contacts.apk "ORIGIN_APP_PATH"Contacts.apk" );
